git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@848 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ admin/svn2package.sh
@@ -10,7 +10,7 @@
 ##########################
 
 function usage() {
-    echo "Usage: $0 [-b|--branch <svn-branch>] [-pdf] [-qch] [packagename]"
+    echo "Usage: $0 [-b|--branch <svn-branch>] [-s|--suffix <suffix>] [-html] [-pdf] [-qch] [packagename]"
     exit 1
 }
 
@@ -83,7 +83,7 @@ function cleanQwt {
         sed -i -e '/#warning/d' $SRCFILE 
     done 
 
-    sed -i -e "s/\$\$QWT_VERSION-svn/\$\$QWT_VERSION-rc2/" qwtconfig.pri 
+    sed -i -e "s/\$\$QWT_VERSION-svn/\$\$QWT_VERSION-$SUFFIX/" qwtconfig.pri 
 
     cd - > /dev/null
 }
@@ -179,7 +179,7 @@ function prepare4Win {
         exit $?
     fi
 
-    rm -r doc/man 
+    rm -rf doc/man 2> /dev/null
 
     # win files, but not uptodate
 
@@ -220,10 +220,12 @@ function prepare4Unix {
 QWTDIR=
 SVNDIR=trunk
 BRANCH=qwt
+SUFFIX=
 VERSION=
+GENERATE_DOC=0
 GENERATE_PDF=0
-GENERATE_QCH=1
-GENERATE_MAN=1
+GENERATE_QCH=0
+GENERATE_MAN=0
 
 while [ $# -gt 0 ] ; do
     case "$1" in
@@ -231,10 +233,14 @@ while [ $# -gt 0 ] ; do
             usage; exit 1 ;;
         -b|--branch)
             shift; SVNDIR=branches; BRANCH=$1; shift;;
+        -s|--suffix)
+            shift; SUFFIX=$1; shift;;
+        -html)
+            GENERATE_DOC=1; shift;;
         -pdf)
-            GENERATE_PDF=1; shift;;
+            GENERATE_DOC=1; GENERATE_PDF=1; shift;;
         -qch)
-            GENERATE_QCH=1; shift;;
+            GENERATE_DOC=1; GENERATE_QCH=1; shift;;
         *) 
             QWTDIR=qwt-$1 ; VERSION=$1; shift;;
     esac
@@ -246,6 +252,11 @@ then
     exit 2 
 fi
 
+if [ "$SUFFIX" != "" ]
+then
+    QWTDIR=$QWTDIR-$SUFFIX
+fi
+
 TMPDIR=/tmp/$QWTDIR-tmp
 
 echo -n "checkout to $TMPDIR ... "
@@ -253,18 +264,21 @@ checkoutQwt $SVNDIR $BRANCH $TMPDIR
 cleanQwt $TMPDIR
 echo done
 
-echo -n "generate documentation ... "
-createDocs $TMPDIR/doc
-
-if [ $GENERATE_PDF -ne 0 ]
+if [ $GENERATE_DOC -ne 0 ]
 then
-    mv $TMPDIR/doc/pdf/qwtdoc-$VERSION.pdf $QWTDIR.pdf
-    rmdir $TMPDIR/doc/pdf
-fi
+	echo -n "generate documentation ... "
+	createDocs $TMPDIR/doc
 
-if [ $GENERATE_QCH -ne 0 ]
-then
-    mv $TMPDIR/doc/html/qwtdoc.qch $QWTDIR.qch
+	if [ $GENERATE_PDF -ne 0 ]
+	then
+    	mv $TMPDIR/doc/pdf/qwtdoc-$VERSION.pdf $QWTDIR.pdf
+    	rmdir $TMPDIR/doc/pdf
+	fi
+
+	if [ $GENERATE_QCH -ne 0 ]
+	then
+    	mv $TMPDIR/doc/html/qwtdoc.qch $QWTDIR.qch
+	fi
 fi
 
 echo done
